+++ includes/class-wootheme-testimonials-to-testimonials-settings.php
@@ -68,7 +68,7 @@ class Wootheme_Testimonials_to_Testimonials_Settings extends Aihrus_Settings {
 
 
 	public static function admin_menu() {
-		self::$admin_page =  add_submenu_page( 'edit.php?post_type=' . Testimonials_Widget::PT, esc_html__( 'WooTheme Testimonials to Testimonials Settings', 'wootheme-testimonials-to-testimonials' ), esc_html__( 'WTT Settings', 'wootheme-testimonials-to-testimonials' ), 'manage_options', self::ID, array( __CLASS__, 'display_page' ) );
+		self::$admin_page = add_submenu_page( 'edit.php?post_type=' . Testimonials_Widget::PT, esc_html__( 'WooTheme Testimonials to Testimonials Settings', 'wootheme-testimonials-to-testimonials' ), esc_html__( 'WTT Settings', 'wootheme-testimonials-to-testimonials' ), 'manage_options', self::ID, array( __CLASS__, 'display_page' ) );
 
 		add_action( 'admin_print_scripts-' . self::$admin_page, array( __CLASS__, 'scripts' ) );
 		add_action( 'admin_print_styles-' . self::$admin_page, array( __CLASS__, 'styles' ) );