+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master
 
+## 1.1.1
+* Coding standards update
+* Require Testimonials by Aihrus 2.18.1
+* Revise Plugin Name: Testimonials – WooTheme Testimonials Migrator
+
 ## 1.1.0
 * Move ci to tests
 * Move CSS to assets